HW05-0036453288/src/main/java/hr/fer/zemris/java/tecaj_2/jcomp/impl/instructions/InstrDecrement.java
package hr.fer.zemris.java.tecaj_2.jcomp.impl.instructions;

import java.util.List;

import hr.fer.zemris.java.tecaj_2.jcomp.Computer;
import hr.fer.zemris.java.tecaj_2.jcomp.Instruction;
import hr.fer.zemris.java.tecaj_2.jcomp.InstructionArgument;

/**
 * Razred <code>InstrDecrement</code> omogućava umanjivanje sadržaja
 * registra za jedan.
 * 
 * <pre>
 *   decrement r1;
 * </pre>
 * 
 * @author Igor Smolkovič
 * @version 1.0
 *
 */
public class InstrDecrement implements Instruction {
	/**
	 * Indeks registra.
	 */
	private int registerIndex;
	
	
	/**
	 * Konstruktor razreda <code>InstrDecrement</code> prima listu argumenata.
	 * 
	 * @param arguments Lista argumenata.
	 * @throws IllegalArgumentException ako lista ne sadrži 1 element ili jedini 
	 * 			element nije registar. 
	 */
	public InstrDecrement(List<InstructionArgument> arguments) {
		if(arguments.size() != 1) {
			throw new IllegalArgumentException("Expected 1 arguments");
		}
		if(!arguments.get(0).isRegister()) {
			throw new IllegalArgumentException("Type mismatch for argument 0!");
		}
		
		this.registerIndex = ((Integer)arguments.get(0).getValue()).intValue();
	}
	
	
	/**
	 * Metoda koja pokreće izvršavanje instrukcije decrement.
	 * 
	 * @param computer Referenca na računalo.
	 * @return false jer nije potrebno zaustaviti rad procesora.
	 */
	@Override
	public boolean execute(Computer computer) {
		Object value = computer.getRegisters().getRegisterValue(registerIndex);
		computer.getRegisters().setRegisterValue(
				registerIndex, 
				((Integer)value).intValue() - 1
		);
		
		return false;
	}

}
